feat(Vim): Set incremental search
This setting enables the incremental search: when searching with "/", Vim will search at each caracter typed, and not wait for "Enter" to be pressed.
This allows useage of the search for navigation as you can see when you get where you want and you don't necessarily have to type everything.