feat(Vim): Indentation handling
General use of spaces rather than tabs.
I've also activated autoindent so new lines are indented somewhat correctly (it's not perfect it, but better than nothing).
The default width of a tab being set to 8 spaces, this isn't compatible with Python which expects 4, so those settings are specified specifically for Python files.