feat(Vim): Autoinstallation of Vim Plug
I've chosen to use Vim Plug (https://github.com/junegunn/vim-plug) as a
plugin manager. The reason for that is that I find it nice that it is
only one file, and it seems like a standard in the Vim community.
The code comes from the Github page of Vim plug linked above. It checks
for Vim plug's presence and uses curl to get it if not found. The Vim-plug file is put in ~/.vim/autoload/plug.vim