chore(Vim): Definition of the mapleader & reorder
I've decided to use the Space bar as leader key. It is easily accessible
from both hands and I don't use it for any other particular action, so I
seemed like a good choice.
I've also reordered the sections of my Vimrc as I'll have keymappings
for plugins, so I think it makes more sense to have the plugins detailed
first and the keymaps after rather than the opposite.