feat(Vim): Airline as a statusline
I'll use the airline plugin as my statusline. It shows everything that I
wanted almost out of the box, and more (active branch name, active Vim
mode, etc). It needs a bit of configuration however to show a tabline on
top of the screen, showing open buffers. It also needs to be told to use
powerline fonts to render properly, and finally I've activated the poetv
extension so it shows the active venv when using a python environment.