feat(Tmux): First tmux.conf file
I've been frustrated when working with Vim and having to interact with
the terminal. I found the usage of Vim commands forwarded to the
terminal to be limiting, and having to leave Vim to do stuff in the
terminal and coming back was a chore as well. Having terminal panes, or
workspaces was also a hassle. After discussing with colleagues, I've
decided to try one of their recommended solutions: using Tmux.
This first conf file is almost a copy of Remy Frenoy's conf file, which
can be found here:
https://github.com/rfrenoy/dotfiles/blob/master/.tmux.conf