fix(Tmux): Fix for Vim's colorscheme
Previously, Vim's Gruvbox colorscheme was not rendered correctly when in
Tmux: the result was different (and considerably worse) then when Vim
ran outside of Tmux. Using the ressources linked below, I noticed that
the $TERM value was different when in Tmux and when outside of Tmux.
This line of conf corrects this, allowing Vim to exhibit the same
behavior inside and outside of Tmux.
https://fossies.org/linux/tmux/example_tmux.conf
https://stackoverflow.com/questions/15375992/vim-difference-between-t-co-256-and-term-xterm-256color-in-conjunction-with-tmu/15378816#15378816
https://unix.stackexchange.com/questions/348771/why-do-vim-colors-look-different-inside-and-outside-of-tmux