chore(Tmux): Adjusting the vertical resize
When Adjusting the  size of 2 horizontal panes (i.e. vertical line
splits the screen in two), it is OK to move the dividing line by 5
columns. However, when the screen is split in 2 vertical panes (i.e. a
hohorizontal line splits the screen in two), moving this line up or down
by 5 lines was a bit too much on my screen. An increment of 3 will be
enough.