feat(powerline): Addition of the config file for powerline
This is required to set the default theme of the shell to left_only.
The default shell tries to display some information on the right side of
the terminal, but this is not possible with bash. Thus, with the
default configuration, I was missing the branch information.
This file is currently the default one, with only that parameter
changed. I'll try to slim it down to only what's needed in next commits.