feat(Tmux): Hot reload of Tmux conf & 1-index for windows
Following my reading of the "tmux 2" book, I've discovered those 2
quality of life changes.
Hot reload of the conf allows for quicker iteration and testing.
Using a window indexing starting at 1 allows for an easier switch
between windows with the shortcut "prefix + number", as the "0" is far
away and completely to the right while the rest are on the left. Now it
makes more sense as the first 9 windows will "follow" each other.